Another attempt at fixing this for both of us.

+++ mcon/io.c
@@ -148,8 +148,8 @@ int mongo_io_wait_with_timeout(int sock, int to, char **error_message)
 			*error_message = malloc(256);
 			snprintf(
 				*error_message, 256,
-				"cursor timed out (timeout: %d, time left: %ld:%d, status: %d)",
-				to, timeout.tv_sec, timeout.tv_usec, status
+				"cursor timed out (timeout: %d, time left: %ld:%ld, status: %d)",
+				to, (long) timeout.tv_sec, (long) timeout.tv_usec, status
 			);
 			return 80;
 		}
